another round of unsigned ints, mostly warnings this time, refs #2495
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21202 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/duarouter/duarouter_main.cpp
@@ -106,7 +106,7 @@ computeRoutes(RONet& net, ROLoader& loader, OptionsCont& oc) {
     // prepare the output
     const std::string& filename = oc.getString("output-file");
     std::string altFilename = filename + ".alt";
-    const int len = filename.length();
+    const int len = (int)filename.length();
     if (len > 4 && filename.substr(len - 4) == ".xml") {
         altFilename = filename.substr(0, len - 4) + ".alt.xml";
     } else if (len > 4 && filename.substr(len - 4) == ".sbx") {
