+++ sumo/src/netimport/vissim/tempstructs/NIVissimNodeCluster.cpp
@@ -114,7 +114,7 @@ NIVissimNodeCluster::dictionary(int id) {
 
 int
 NIVissimNodeCluster::contSize() {
-    return myDict.size();
+    return (int)myDict.size();
 }
