+++ sumo/src/router/RORouteDef.cpp
@@ -176,7 +176,7 @@ RORouteDef::repairCurrentRoute(SUMOAbstractRouter<ROEdge, ROVehicle>& router,
     MsgHandler* mh = (OptionsCont::getOptions().getBool("ignore-errors") ?
                       MsgHandler::getWarningInstance() : MsgHandler::getErrorInstance());
     ConstROEdgeVector mandatory;
-    const int initialSize = oldEdges.size();
+    const int initialSize = (int)oldEdges.size();
     if (initialSize == 1) {
         if (myUsingJTRR) {
             /// only ROJTRRouter is supposed to handle this type of input
